refactor(client): reduce imports in entity edit & view pages
The import/max-dependencies lint rule (threshold 10) flagged the entity
edit/view route pages, which had accreted shared scaffolding and wide
form bodies past the point where they should be decomposed. Group the
shared pieces behind cohesive barrels and componentize the heaviest
pages so each file pulls from fewer modules — no behavior change.

- Add components/editPage and components/infoCard barrels for the
  edit-page shell and view-page info sections; re-export queryKeys from
  @/utils and ResourceLinksPicker from formFields so single-use imports
  collapse into existing ones.
- Extract the topics edit form into a route-private -TopicForm, and split
  domains edit into -NewDomainForm + -ExistingDomainEditor with a -tabs
  barrel, leaving both route files as thin dispatchers.
- All five target files now sit at <=10 value imports with no
  eslint-disable; typecheck and the client test suite stay green.

Skipped the issue's "namespace the API imports" step: each page already
imports its per-entity API functions from a single @/utils statement, so
it yields no reduction.

Fixes #305
