+++ packages/client/src/components/formFields/index.ts
@@ -1,1 +1,2 @@
+export { ResourceLinksPicker } from "@/components/tasks/ResourceLinksPicker";
 export { useAppForm } from "@/hooks/useAppForm";
